Test house polygon definitions
